refactor: fix cohesion violations — inline tui_history, merge headless_sink
tui_history.rs (38 lines, only consumer: tui_context.rs) inlined.
headless_sink.rs (134 lines, only consumer: headless.rs) merged.

Both were split purely for line count, not because they had
independent responsibilities. Applying the new convention from CLAUDE.md.
