+++ koda-cli/src/main.rs
@@ -5,7 +5,6 @@
 mod completer;
 mod diff_render;
 mod headless;
-mod headless_sink;
 mod highlight;
 mod input;
 mod md_render;
@@ -18,7 +17,6 @@ mod tool_history;
 mod tui_app;
 mod tui_commands;
 mod tui_context;
-mod tui_history;
 mod tui_output;
 mod tui_render;
 mod tui_types;
